Way too much going on in this one...
Created the controller and the get seems to be working. Need to add the post so that I can add entries and make sure.

Created a dto package where the objects that are used to interact with the REST interface will live.

Created a Contact Service that will act as the business logic layer and a seam in between the database and the entry point to the API.

Added the spring boot dev tools as a dependency so that I could hit the H2 console through that instead. Removed the application.properties entry for it.

Added a test for mapping the text for phonetypes to the enum.

Moved the database objects inside of the db.models package. Removed the scan notation from the application class, it doesn't seem to be needed.